+++ dataland-community-manager/src/test/kotlin/org/dataland/datalandcommunitymanager/utils/TestUtils.kt
@@ -1,0 +1,21 @@
+package org.dataland.datalandcommunitymanager.utils
+
+import org.dataland.keycloakAdapter.auth.DatalandRealmRole
+import org.dataland.keycloakAdapter.utils.AuthenticationMock
+import org.mockito.Mockito.mock
+import org.mockito.Mockito.`when`
+import org.springframework.security.core.context.SecurityContext
+import org.springframework.security.core.context.SecurityContextHolder
+
+class TestUtils {
+    fun mockSecurityContext() {
+        val mockAuthentication = AuthenticationMock.mockJwtAuthentication(
+            "mocked_uploader",
+            "dummy-id",
+            setOf(DatalandRealmRole.ROLE_PREMIUM_USER),
+        )
+        val mockSecurityContext = mock(SecurityContext::class.java)
+        `when`(mockSecurityContext.authentication).thenReturn(mockAuthentication)
+        SecurityContextHolder.setContext(mockSecurityContext)
+    }
+}
